COMP1531/project/requirements.txt
apipkg==1.5
appdirs==1.4.4
astroid==2.4.2
attrs==20.2.0
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
coverage==5.2.1
distlib==0.3.1
dnspython==2.0.0
email-validator==1.1.1
execnet==1.7.1
filelock==3.0.12
Flask==1.1.2
Flask-Cors==3.0.9
hypothesis==5.24.3
idna==2.10
importlib-metadata==2.0.0
iniconfig==1.1.1
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.5.0
mypy==0.782
mypy-extensions==0.4.3
packaging==20.4
Pillow==8.0.1
pluggy==0.13.1
py==1.9.0
pycodestyle==2.6.0
PyJWT==1.7.1
pylint==2.5.3
pylint-venv==2.1.1
pyparsing==2.4.7
pytest==6.0.1
pytest-env==0.6.2
pytest-forked==1.3.0
requests==2.24.0
rope==0.18.0
six==1.15.0
sortedcontainers==2.2.2
toml==0.10.1
typed-ast==1.4.1
typing-extensions==3.7.4.3
urllib3==1.25.10
virtualenv==20.0.31
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.3.1
